Change Type of Integer to Byte Vector Converter to Unsigned Int 8

+++ test/testing_utilities.cpp
@@ -7,7 +7,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 
-std::vector<std::byte> intVec2ByteVec(const std::vector<int>& intVec) {
+std::vector<std::byte> intVec2ByteVec(const std::vector<uint8_t>& intVec) {
     std::vector<std::byte> byteVec(intVec.size());
     for (size_t i = 0; i < intVec.size(); ++i)
         byteVec[i] = static_cast<std::byte>(intVec[i]);
